Fix out-of-bounds error in encoding (#21)
## Overview

Previously, the generated `encode()` function was panicking for an
`array out-of-bounds access` error. This was because the loops generated
for setting `finalEncoded` did not increment `j`:

<img width="572" alt="image"
src="https://github.com/celestiaorg/protobuf3-solidity/assets/22297592/aada53a4-0b88-4078-9489-51ec3cfd202b">

This PR fixes this. Tested it and we don't encounter that error anymore.


## Checklist

- [x] New and updated code has appropriate documentation
- [x] New and updated code has new and/or updated testing
- [x] Required CI checks are passing
- [x] Visual proof for any user facing features like CLI or
documentation updates
- [x] Linked issues closed with keywords